refactor io module for more features
